Replace hard coded fonts with dynamic fonts
